requirements.txt
Django==5.2
django-model-utils==5.0.0
django-redis==5.4.0
django-thumbnails==0.8.0

psycopg2-binary==2.9.10
redis==5.2.1